# docs/api/openapi-spec.yaml
openapi: 3.0.3
info:
  title: AI Trading Platform API
  description: |
    Comprehensive REST API for the AI Trading Platform providing access to trading signals, 
    portfolio management, risk monitoring, and model serving capabilities.
    
    ## Features
    - Multi-exchange trading support (Robinhood, OANDA, Coinbase)
    - AI-powered trading signals with CNN+LSTM models
    - Real-time market data streaming via WebSocket
    - Portfolio optimization and risk management
    - Model interpretability and explainability
    
    ## Authentication
    All endpoints require JWT authentication. Obtain a token via `/auth/login`.
    
    ## Rate Limits
    - Free Tier: 100 requests/hour, 5 signals/day
    - Pro Tier: 1000 requests/hour, 100 signals/day
    - Enterprise: Custom limits
    
  version: 1.0.0
  contact:
    name: AI Trading Platform Support
    email: api-support@ai-trading-platform.com
    url: https://docs.ai-trading-platform.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.ai-trading-platform.com
    description: Production server
  - url: https://sandbox-api.ai-trading-platform.com
    description: Sandbox server
  - url: http://localhost:8000
    description: Local development server
security:
  - BearerAuth: []
paths:
...
  /api/v1/portfolio/rebalance:
    post:
      tags:
        - Portfolio
      summary: Rebalance portfolio
      description: Trigger portfolio rebalancing with specified strategy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RebalanceRequest'
      responses:
        '202':
          description: Rebalancing initiated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RebalanceResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    RebalanceRequest:
      type: object
      required:
        - strategy
      properties:
        strategy:
          type: string
          enum: [equal_weight, risk_parity, momentum]
        constraints:
          type: object
          properties:
            max_position_size:
              type: number
            min_cash_ratio:
              type: number
    RebalanceResponse:
      type: object
      properties:
        job_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending, running, completed, failed]
        estimated_completion:
          type: string
          format: date-time
...